moving todoitem and todolist classes to righr place
